backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-url-filter
djangorestframework-simplejwt
gunicorn
scrape-schema-recipe
requests